Make check for require(...) a little more explicit
Avoids type assertions. On the one hand, this seems pretty verbose, but
on the other, it really lays out the assumptions you're making about
what kinda node you're working with. Could be helpful in the future to
look for areas to extend support for a type of import or areas where I'm
missing imports.